@@ -17,41 +17,25 @@ use Symfony\Component\Serializer\Annotation as Serializer;
 
 class NodesSourcesHead implements NodesSourcesHeadInterface
 {
-    /**
-     * @var HandlerFactoryInterface
-     * @Serializer\Ignore
-     */
+    #[Serializer\Ignore]
     protected HandlerFactoryInterface $handlerFactory;
 
-    /**
-     * @var array|null
-     * @Serializer\Ignore
-     */
+    #[Serializer\Ignore]
     protected ?array $seo = null;
 
-    /**
-     * @var NodesSources|null
-     * @Serializer\Ignore
-     */
+    #[Serializer\Ignore]
     protected ?NodesSources $nodesSource;
-    /**
-     * @var Settings
-     * @Serializer\Ignore
-     */
+
+    #[Serializer\Ignore]
     protected Settings $settingsBag;
-    /**
-     * @var UrlGeneratorInterface
-     * @Serializer\Ignore
-     */
+
+    #[Serializer\Ignore]
     protected UrlGeneratorInterface $urlGenerator;
-    /**
-     * @var NodeSourceApi
-     * @Serializer\Ignore
-     */
+
+    #[Serializer\Ignore]
     protected NodeSourceApi $nodeSourceApi;
-    /**
-     * @Serializer\Ignore
-     */
+
+    #[Serializer\Ignore]
     protected TranslationInterface $defaultTranslation;
 
     /**
@@ -80,8 +64,8 @@ class NodesSourcesHead implements NodesSourcesHeadInterface
 
     /**
      * @return string|null
-     * @Serializer\Groups({"web_response", "nodes_sources_single", "walker"})
      */
+    #[Serializer\Groups(["web_response", "nodes_sources_single", "walker"])]
     public function getGoogleAnalytics(): ?string
     {
         return $this->settingsBag->get('universal_analytics_id', null) ?? null;
@@ -89,8 +73,8 @@ class NodesSourcesHead implements NodesSourcesHeadInterface
 
     /**
      * @return string|null
-     * @Serializer\Groups({"web_response", "nodes_sources_single", "walker"})
      */
+    #[Serializer\Groups(["web_response", "nodes_sources_single", "walker"])]
     public function getGoogleTagManager(): ?string
     {
         return $this->settingsBag->get('google_tag_manager_id', null) ?? null;
@@ -98,8 +82,8 @@ class NodesSourcesHead implements NodesSourcesHeadInterface
 
     /**
      * @return string|null
-     * @Serializer\Groups({"web_response", "nodes_sources_single", "walker"})
      */
+    #[Serializer\Groups(["web_response", "nodes_sources_single", "walker"])]
     public function getMatomoUrl(): ?string
     {
         return $this->settingsBag->get('matomo_url', null) ?? null;
@@ -107,8 +91,8 @@ class NodesSourcesHead implements NodesSourcesHeadInterface
 
     /**
      * @return string|null
-     * @Serializer\Groups({"web_response", "nodes_sources_single", "walker"})
      */
+    #[Serializer\Groups(["web_response", "nodes_sources_single", "walker"])]
     public function getMatomoSiteId(): ?string
     {
         return $this->settingsBag->get('matomo_site_id', null) ?? null;
@@ -116,8 +100,8 @@ class NodesSourcesHead implements NodesSourcesHeadInterface
 
     /**
      * @return string|null
-     * @Serializer\Groups({"web_response", "nodes_sources_single", "walker"})
      */
+    #[Serializer\Groups(["web_response", "nodes_sources_single", "walker"])]
     public function getSiteName(): ?string
     {
         // site_name
@@ -144,8 +128,8 @@ class NodesSourcesHead implements NodesSourcesHeadInterface
 
     /**
      * @return string|null
-     * @Serializer\Groups({"web_response", "nodes_sources_single", "walker"})
      */
+    #[Serializer\Groups(["web_response", "nodes_sources_single", "walker"])]
     public function getMetaTitle(): ?string
     {
         if (null === $this->seo) {
@@ -157,8 +141,8 @@ class NodesSourcesHead implements NodesSourcesHeadInterface
 
     /**
      * @return string|null
-     * @Serializer\Groups({"web_response", "nodes_sources_single", "walker"})
      */
+    #[Serializer\Groups(["web_response", "nodes_sources_single", "walker"])]
     public function getMetaDescription(): ?string
     {
         if (null === $this->seo) {
@@ -170,8 +154,8 @@ class NodesSourcesHead implements NodesSourcesHeadInterface
 
     /**
      * @return string|null
-     * @Serializer\Groups({"web_response", "nodes_sources_single", "walker"})
      */
+    #[Serializer\Groups(["web_response", "nodes_sources_single", "walker"])]
     public function getPolicyUrl(): ?string
     {
         $translation = $this->getTranslation();
@@ -196,8 +180,8 @@ class NodesSourcesHead implements NodesSourcesHeadInterface
 
     /**
      * @return string|null
-     * @Serializer\Groups({"web_response", "nodes_sources_single", "walker"})
      */
+    #[Serializer\Groups(["web_response", "nodes_sources_single", "walker"])]
     public function getMainColor(): ?string
     {
         return $this->settingsBag->get('main_color', null) ?? null;
@@ -205,8 +189,8 @@ class NodesSourcesHead implements NodesSourcesHeadInterface
 
     /**
      * @return string|null
-     * @Serializer\Groups({"web_response", "nodes_sources_single", "walker"})
      */
+    #[Serializer\Groups(["web_response", "nodes_sources_single", "walker"])]
     public function getFacebookUrl(): ?string
     {
         return $this->settingsBag->get('facebook_url', null) ?? null;
@@ -214,8 +198,8 @@ class NodesSourcesHead implements NodesSourcesHeadInterface
 
     /**
      * @return string|null
-     * @Serializer\Groups({"web_response", "nodes_sources_single", "walker"})
      */
+    #[Serializer\Groups(["web_response", "nodes_sources_single", "walker"])]
     public function getInstagramUrl(): ?string
     {
         return $this->settingsBag->get('instagram_url', null) ?? null;
@@ -223,8 +207,8 @@ class NodesSourcesHead implements NodesSourcesHeadInterface
 
     /**
      * @return string|null
-     * @Serializer\Groups({"web_response", "nodes_sources_single", "walker"})
      */
+    #[Serializer\Groups(["web_response", "nodes_sources_single", "walker"])]
     public function getTwitterUrl(): ?string
     {
         return $this->settingsBag->get('twitter_url', null) ?? null;
@@ -232,8 +216,8 @@ class NodesSourcesHead implements NodesSourcesHeadInterface
 
     /**
      * @return string|null
-     * @Serializer\Groups({"web_response", "nodes_sources_single", "walker"})
      */
+    #[Serializer\Groups(["web_response", "nodes_sources_single", "walker"])]
     public function getYoutubeUrl(): ?string
     {
         return $this->settingsBag->get('youtube_url', null) ?? null;
@@ -241,8 +225,8 @@ class NodesSourcesHead implements NodesSourcesHeadInterface
 
     /**
      * @return string|null
-     * @Serializer\Groups({"nodes_sources_single", "walker"})
      */
+    #[Serializer\Groups(["nodes_sources_single", "walker"])]
     public function getLinkedinUrl(): ?string
     {
         return $this->settingsBag->get('linkedin_url', null) ?? null;
@@ -250,8 +234,8 @@ class NodesSourcesHead implements NodesSourcesHeadInterface
 
     /**
      * @return string|null
-     * @Serializer\Groups({"web_response", "nodes_sources_single", "walker"})
      */
+    #[Serializer\Groups(["web_response", "nodes_sources_single", "walker"])]
     public function getHomePageUrl(): ?string
     {
         $homePage = $this->getHomePage();
@@ -265,8 +249,8 @@ class NodesSourcesHead implements NodesSourcesHeadInterface
 
     /**
      * @return DocumentInterface|null
-     * @Serializer\Groups({"web_response", "nodes_sources_single"})
      */
+    #[Serializer\Groups(["web_response", "nodes_sources_single"])]
     public function getShareImage(): ?DocumentInterface
     {
         if (
@@ -288,8 +272,8 @@ class NodesSourcesHead implements NodesSourcesHeadInterface
 
     /**
      * @return TranslationInterface
-     * @Serializer\Ignore()
      */
+    #[Serializer\Ignore()]
     public function getTranslation(): TranslationInterface
     {
         if (null !== $this->nodesSource) {
@@ -300,8 +284,8 @@ class NodesSourcesHead implements NodesSourcesHeadInterface
 
     /**
      * @return NodesSources|null
-     * @Serializer\Ignore()
      */
+    #[Serializer\Ignore()]
     private function getHomePage(): ?NodesSources
     {
         return $this->nodeSourceApi->getOneBy([
