@@ -4,16 +4,18 @@ declare(strict_types=1);
 
 namespace RZ\Roadiz\CoreBundle\Entity;
 
-use ApiPlatform\Core\Annotation\ApiFilter;
+use ApiPlatform\Metadata\ApiFilter;
+use ApiPlatform\Core\Bridge\Doctrine\Orm\Filter as BaseFilter;
+use ApiPlatform\Core\Serializer\Filter\PropertyFilter;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\Mapping as ORM;
 use JMS\Serializer\Annotation as Serializer;
+use RZ\Roadiz\Core\AbstractEntities\AbstractPositioned;
 use RZ\Roadiz\CoreBundle\Model\AttributableInterface;
 use RZ\Roadiz\CoreBundle\Model\AttributeValueInterface;
 use RZ\Roadiz\CoreBundle\Model\AttributeValueTrait;
 use RZ\Roadiz\CoreBundle\Model\AttributeValueTranslationInterface;
-use ApiPlatform\Core\Bridge\Doctrine\Orm\Filter as BaseFilter;
-use RZ\Roadiz\Core\AbstractEntities\AbstractPositioned;
+use Symfony\Component\Serializer\Annotation as SymfonySerializer;
 
 /**
  * @package RZ\Roadiz\CoreBundle\Entity
@@ -23,6 +25,7 @@ use RZ\Roadiz\Core\AbstractEntities\AbstractPositioned;
  * })
  * @ORM\HasLifecycleCallbacks
  */
+#[ApiFilter(PropertyFilter::class)]
 class AttributeValue extends AbstractPositioned implements AttributeValueInterface
 {
     use AttributeValueTrait;
@@ -31,15 +34,18 @@ class AttributeValue extends AbstractPositioned implements AttributeValueInterfa
      * @var Node|null
      * @ORM\ManyToOne(targetEntity="RZ\Roadiz\CoreBundle\Entity\Node", inversedBy="attributeValues")
      * @ORM\JoinColumn(name="node_id", onDelete="CASCADE")
-     * @ApiFilter(BaseFilter\SearchFilter::class, properties={
-     *     "node.id": "exact",
-     *     "node.nodeName": "exact"
-     * })
-     * @ApiFilter(BaseFilter\BooleanFilter::class, properties={
-     *     "node.visible"
-     * })
-     * @Serializer\Exclude
+     * @Serializer\Groups({"attribute_node"})
+     * @SymfonySerializer\Groups({"attribute_node"})
+     * @SymfonySerializer\MaxDepth(1)
      */
+    #[ApiFilter(BaseFilter\SearchFilter::class, properties: [
+        "node" => "exact",
+        "node.id" => "exact",
+        "node.nodeName" => "exact"
+    ])]
+    #[ApiFilter(BaseFilter\BooleanFilter::class, properties: [
+        "node.visible"
+    ])]
     protected ?Node $node = null;
 
     public function __construct()
