@@ -4,7 +4,8 @@ declare(strict_types=1);
 
 namespace RZ\Roadiz\CoreBundle\Entity;
 
-use ApiPlatform\Core\Annotation\ApiFilter;
+use ApiPlatform\Core\Serializer\Filter\PropertyFilter;
+use ApiPlatform\Metadata\ApiFilter;
 use ApiPlatform\Core\Bridge\Doctrine\Orm\Filter as BaseFilter;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
@@ -31,26 +32,26 @@ use Symfony\Component\Validator\Constraints as Assert;
  *     @ORM\Index(columns={"available", "locale"}),
  *     @ORM\Index(columns={"available", "override_locale"})
  * })
- * @ApiFilter(\ApiPlatform\Core\Serializer\Filter\PropertyFilter::class)
- * @ApiFilter(BaseFilter\OrderFilter::class, properties={
- *     "createdAt",
- *     "updatedAt",
- *     "locale",
- *     "available",
- *     "defaultTranslation"
- * })
- * @ApiFilter(BaseFilter\BooleanFilter::class, properties={
- *     "available",
- *     "defaultTranslation"
- * })
- * @ApiFilter(BaseFilter\SearchFilter::class, properties={
- *     "locale": "exact",
- *     "name": "exact"
- * })
  * @UniqueEntity(fields={"name"})
  * @UniqueEntity(fields={"locale"})
  * @UniqueEntity(fields={"overrideLocale"})
  */
+#[ApiFilter(PropertyFilter::class)]
+#[ApiFilter(BaseFilter\OrderFilter::class, properties: [
+    "createdAt",
+    "updatedAt",
+    "locale",
+    "available",
+    "defaultTranslation"
+])]
+#[ApiFilter(BaseFilter\BooleanFilter::class, properties: [
+    "available",
+    "defaultTranslation"
+])]
+#[ApiFilter(BaseFilter\SearchFilter::class, properties: [
+    "locale" => "exact",
+    "name" => "exact"
+])]
 class Translation extends AbstractDateTimed implements TranslationInterface
 {
     /**
@@ -550,7 +551,7 @@ class Translation extends AbstractDateTimed implements TranslationInterface
      * @var string
      * @ORM\Column(type="string", unique=true, length=10)
      * @Serializer\Groups({"translation", "document", "nodes_sources", "tag", "attribute", "folder", "log_sources"})
-     * @SymfonySerializer\Groups({"translation", "document", "nodes_sources", "tag", "attribute", "folder", "log_sources"})
+     * @SymfonySerializer\Ignore()
      * @Serializer\Type("string")
      * @Assert\NotBlank()
      * @Assert\NotNull()
@@ -561,7 +562,7 @@ class Translation extends AbstractDateTimed implements TranslationInterface
      * @var string|null
      * @ORM\Column(type="string", name="override_locale", length=10, unique=true, nullable=true)
      * @Serializer\Groups({"translation", "document", "nodes_sources", "tag", "attribute", "folder"})
-     * @SymfonySerializer\Groups({"translation", "document", "nodes_sources", "tag", "attribute", "folder"})
+     * @SymfonySerializer\Ignore()
      * @Serializer\Type("string")
      * @Assert\Length(max=10)
      */
@@ -569,8 +570,8 @@ class Translation extends AbstractDateTimed implements TranslationInterface
     /**
      * @var string
      * @ORM\Column(type="string", unique=true)
-     * @Serializer\Groups({"translation"})
-     * @SymfonySerializer\Groups({"translation"})
+     * @Serializer\Groups({"translation", "translation_base"})
+     * @SymfonySerializer\Groups({"translation", "translation_base"})
      * @Serializer\Type("string")
      * @Assert\NotNull()
      * @Assert\NotBlank()
@@ -580,16 +581,16 @@ class Translation extends AbstractDateTimed implements TranslationInterface
     /**
      * @var bool
      * @ORM\Column(name="default_translation", type="boolean", nullable=false, options={"default" = false})
-     * @Serializer\Groups({"translation"})
-     * @SymfonySerializer\Groups({"translation"})
+     * @Serializer\Groups({"translation", "translation_base"})
+     * @SymfonySerializer\Groups({"translation", "translation_base"})
      * @Serializer\Type("bool")
      */
     private bool $defaultTranslation = false;
     /**
      * @var bool
      * @ORM\Column(type="boolean", nullable=false, options={"default" = true})
-     * @Serializer\Groups({"translation"})
-     * @SymfonySerializer\Groups({"translation"})
+     * @Serializer\Groups({"translation", "translation_base"})
+     * @SymfonySerializer\Groups({"translation", "translation_base"})
      * @Serializer\Type("bool")
      */
     private bool $available = true;
@@ -775,6 +776,8 @@ class Translation extends AbstractDateTimed implements TranslationInterface
      * Get preferred locale between overrideLocale or locale.
      *
      * @return string
+     * @SymfonySerializer\SerializedName("locale")
+     * @SymfonySerializer\Groups({"translation_base"})
      */
     public function getPreferredLocale(): string
     {
